feat: add hello-universe 1.2.0 pack DOC-1138
This PR includes a new version of the Hello Universe pack in the Spectro Cloud community repo. It bumps the version of the Hello Universe image to 1.2.0 and the API & DB versions to 1.1.0.
Hello Universe is a demo web application that helps users learn more about Palette and its features.